rename to state_machine
