+++ state-chain/pallets/cf-elections/src/electoral_systems/state_machine.rs
@@ -1,4 +1,4 @@
 pub mod consensus;
 pub mod core;
-pub mod state_machine2;
+pub mod state_machine;
 pub mod state_machine_es;
