+++ state-chain/pallets/cf-elections/src/electoral_systems/state_machine/state_machine.rs
