+++ state-chain/pallets/cf-elections/src/electoral_systems/state_machine/state_machine_es.rs
@@ -12,7 +12,7 @@ use crate::{
 use super::{
 	consensus::{ConsensusMechanism, Threshold},
 	core::{Indexed, MultiIndexAndValue, Validate},
-	state_machine2::StateMachine,
+	state_machine::StateMachine,
 };
 
 pub trait IntoResult {
